@@ -1,15 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
+
 import { CounterContextProvider } from './context/CounterContext';
 import './styles/global-styles.css';
 
 import { Home } from './templates/Home';
-
-ReactDOM.render(
+const rootElement = document.getElementById('root');
+const root = createRoot(rootElement);
+root.render(
   <React.StrictMode>
     <CounterContextProvider>
       <Home />
     </CounterContextProvider>
   </React.StrictMode>,
-  document.getElementById('root'),
 );
