@@ -1,5 +1,7 @@
 import P from 'prop-types';
-import { createContext, useContext, useState } from 'react';
+import { createContext, useContext, useReducer, useRef, useState } from 'react';
+import { buildActions } from './build-actions';
+import { reducer } from './reducer.';
 
 export const initialState = {
   counter: 0,
@@ -7,9 +9,16 @@ export const initialState = {
 };
 
 const Context = createContext();
+
 export const CounterContextProvider = ({ children }) => {
-  const [state, dispatch] = useState(initialState);
-  return <Context.Provider value={[state, dispatch]}>{children}</Context.Provider>;
+  const [state, dispatch] = useReducer(reducer, initialState);
+  const actions = useRef(buildActions(dispatch));
+
+  return (
+    <Context.Provider value={[state, actions.current]}>
+      {children}
+    </Context.Provider>
+  );
 };
 
 CounterContextProvider.propTypes = {
@@ -20,7 +29,9 @@ export const useCounterContext = () => {
   const context = useContext(Context);
 
   if (typeof context === 'undefined') {
-    throw new Error('You have to useCounterContext inside <CounterContextProvider />');
+    throw new Error(
+      'You have to use useCounterContext inside <CounterContextProvider />',
+    );
   }
 
   return [...context];
